Comment, rewrite and remove examples
