+++ examples/empty.c
@@ -1,1 +1,7 @@
+/*
+ * empty.c: Do nothing
+ *
+ * Usage: ./empty.sh
+ */
+
 void main() {}
