+++ examples/fib.c
@@ -1,3 +1,9 @@
+/*
+ * fib.c: Print the first 20 Fibonacci numbers.
+ *
+ * Usage: ./fib.sh
+ */
+
 #include <stdio.h>
 
 int fib(int n) {
