+++ examples/hello.c
@@ -1,3 +1,9 @@
+/*
+ * hello.c: Print "Hello, world"
+ *
+ * Usage: ./hello.sh
+ */
+
 #include <stdio.h>
 
 int main() {
